+++ game/wrenapi.h
@@ -1,7 +1,7 @@
 #pragma once
 
 struct WrenVM *Wren_Init(const char *mainScriptName, const char *constructorStr);
-void Wren_Update(WrenVM *vm, float dt);
+void Wren_Update(WrenVM *vm, double dt);
 void Wren_Draw(struct WrenVM *vm, int w, int h);
 void Wren_Eval(WrenVM *vm, const char *code);
 void Wren_Console(WrenVM *vm, const char *str);
